// client/src/config/manifest.ts
// This will be generated by your Dojo build
// For now, we'll create a mock version

export const manifest = {
  world: {
    address: import.meta.env.VITE_PUBLIC_WORLD_ADDRESS,
    name: "rift_commanders",
  },
  contracts: [
    {
      address: import.meta.env.VITE_PUBLIC_WORLD_ADDRESS,
      name: "rift_commanders",
    },
  ],
};